scoreStationDistancesの展開
